ai-server/requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
tokenizers>=0.13.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
networkx>=3.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Hugging Face Hub
huggingface-hub>=0.15.0

# RAG and AI Dependencies
faiss-cpu>=1.7.0
google-generativeai>=0.3.0

# Jupyter/Notebook Support (for development)
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Visualization (optional)
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pathlib
python-multipart
python-json-logger

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Environment management
python-dotenv>=1.0.0